simulator/requirements.txt
# HeatAQ Simulator Dependencies
mysql-connector-python>=8.0.0
numpy>=1.20.0
pandas>=1.3.0
